requirements.txt
langchain
langchain-community
langchain-openai
streamlit
chromadb
pypdf
python-dotenv
langchain_chroma